Monitor file for changes
Minor fixes to README
Minor style fixes
Stat file every FILE_CHECK_TIMEOUT_SEC seconds for changes

+++ text-pango.h
@@ -47,5 +47,5 @@ struct pango_source {
 
 	bool file_load_failed;
 	time_t file_timestamp;
-	uint64_t file_last_checked;
+	float file_last_checked; // Sec since last checked
 };
